// src/app/model/marker-details.ts
// Generated by https?://quicktype.io

export interface MarkerDetails {
    id?: number;
    name?: string;
    status?: number;
    countryFullName?: string;
    city?: string;
    mainPhotoUrl?: string;
    ratePerHour?: number;
    currency?: number;
    lat?: number;
    lon?: number;
    maxParticipants?: number;
    ratingAvg?: null;
    distance?: number;
    isOn?: boolean;
}